Move onion struct to a new file
