+++ src/node_api.rs
@@ -1,16 +1,14 @@
 use crate::disk::FilesystemLogger;
+use crate::onion::UserOnionMessageContents;
 use crate::{
     BitcoindClient, ChainMonitor, ChannelManager, NetworkGraph, OnionMessenger, PeerManager,
 };
 
 use bitcoin::secp256k1::PublicKey;
-use lightning::onion_message::{
-    CustomOnionMessageContents, Destination, OnionMessageContents, OnionMessagePath,
-};
+use lightning::onion_message::{Destination, OnionMessageContents, OnionMessagePath};
 use lightning::routing::router::DefaultRouter;
 use lightning::routing::scoring::{ProbabilisticScorer, ProbabilisticScoringFeeParameters};
 use lightning::sign::KeysManager;
-use lightning::util::ser::{Writeable, Writer};
 use lightning_persister::FilesystemPersister;
 use std::net::{IpAddr, Ipv4Addr, SocketAddr};
 use std::sync::{Arc, Mutex};
@@ -33,23 +31,6 @@ pub(crate) type P2PGossipSyncType = lightning::routing::gossip::P2PGossipSync<
     Arc<BitcoindClient>,
     Arc<FilesystemLogger>,
 >;
-
-struct UserOnionMessageContents {
-    tlv_type: u64,
-    data: Vec<u8>,
-}
-
-impl CustomOnionMessageContents for UserOnionMessageContents {
-    fn tlv_type(&self) -> u64 {
-        self.tlv_type
-    }
-}
-
-impl Writeable for UserOnionMessageContents {
-    fn write<W: Writer>(&self, w: &mut W) -> Result<(), std::io::Error> {
-        w.write_all(&self.data)
-    }
-}
 
 pub struct Node {
     pub(crate) logger: Arc<FilesystemLogger>,
